refactor: extract charts app from blowcomotion monolith
