+++ charts/templates/chart_import/picker.html
