+++ charts/templates/chart_import/review.html
